temp/requirements-simple.txt
pyautogui>=0.9.53
pytesseract>=0.3.9
pillow>=9.0.0
opencv-python>=4.5.5
numpy>=1.20.0
pdfrw>=0.4.0
torch>=1.10.0
transformers>=4.16.0
pywin32; platform_system=="Windows" 